Update stage_deploy_runner.yml
New version of the ERAS Workflow for the runner.
Changes: The variables were modified to follow a standard, allowing me to use a reduces logic to select the environment setting a SUFFIX to use all the variables of any environment only changing the suffix EX: ERAS_{0}_COSMIC_API_URL should be converted according the suffix in ERAS_DEV_COSMIC_API_URL or ERAS_STAGE_COSMIC_API_URL depending if the environment selected is the develop or the stagging, the environment name are now normalized in a variable avoiding the error on FE and BE docker hub image to have the docker image with the suffix stagging now the variable converts stagging in stage, added a new variable for the branch, the logic of the case fixes the variable in the expected branch for every deployment, having the stage as main, develop and testing as develop and the direct conversion from a github variable is for release, The variable ERAS_RELEASE_BRANCH must be populated before a release with the expected release branch, is the unique manual action before to send a release deploy. this variable is send to the BE and FE sub modules too, so as an additional rule before a release, all the submodules should have a branch release with the same name as the ERAS Project, on the Deploy task an additional option was added to the set of SSH credential for the production environment, our SSH connection is still fixed to use the ubuntu user, we need to evaluate if in a future we are going to set a different user for a different deployments, maybe can improve the security if our working user is not the default user for ubuntu distros.